\name{delete_population}
\alias{delete_population}
\title{Delete a population}
\description{Delete a population from EDP.}
\usage{
delete_population(pop)
}
\arguments{
  \item{pop}{a population from \code{\link{create_population}} or
      \code{\link{population}}}
}
\seealso{
  \code{\link{create_population}}, \code{\link{population}}
}
