\name{edpclient-package}
\alias{edpclient-package}
\alias{edpclient}
\docType{package}
\title{
\packageTitle{edpclient}
}
\description{
\packageDescription{edpclient}
}
\details{
edpclient is the R client for Empirical Data Platform.
}
\author{
\packageAuthor{edpclient}

Maintainer: \packageMaintainer{edpclient}
}
